refactor: move seq2seq * gobot to 'models'
