+++ deeppavlov/models/go_bot/diagram.png
