+++ deeppavlov/models/go_bot/network.py
