+++ deeppavlov/models/go_bot/templates.py
